requirements.txt
# Core dependencies for DOEBI agent infrastructure
# Generated from CLAUDE.md specifications

# Core - Agent orchestration and API
anthropic>=0.18.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Google APIs - For Sheets, Slides, Drive integration
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Environment and Configuration
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# Email
sendgrid>=6.10.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# Slack Integration
slack-sdk>=3.23.0

# Airtable Integration
pyairtable>=2.1.0
